Change SizeValueType definitions to use size_t and ptrdiff_t
This commit enables the use of image sizes, offsets and indexes
larger than 2GB on Win64.

+++ Code/Common/itkIndex.h
@@ -71,7 +71,7 @@ public:
 
   /** Compatible Index and value typedef */
   typedef   Index< VIndexDimension > IndexType;
-  typedef   long                     IndexValueType;
+  typedef   ptrdiff_t                IndexValueType;
 
   /** Get the dimension (size) of the index. */
   static unsigned int GetIndexDimension() { return VIndexDimension; }
